Project PA - 16/06/2021

+++ src/atividade/pa/exibir/LeiaNumero.java
@@ -6,17 +6,18 @@ public class LeiaNumero {
     public void exibirNumero() {
         Scanner scanner = new Scanner(System.in);
 
-        int[] posicaoArray = new int[10];
+        float[] posicaoArray = new float[10];
 
         for (int counter = 0; counter < posicaoArray.length; counter++) {
             System.out.println("Coloque o " + (counter + 1) + " primeiro número: ");
-            posicaoArray[counter] = scanner.nextInt();
+            posicaoArray[counter] = scanner.nextFloat();
 
         }
-
         System.out.println("Os número que você colocou são: ");
-        for (int posicaoArrayLoop : posicaoArray) {
+
+        for (float posicaoArrayLoop : posicaoArray) {
             System.out.println("número: " + posicaoArrayLoop);
         }
+
     }
 }
